+++ docassemble_webapp/docassemble/webapp/config_worker.py
@@ -5,3 +5,5 @@ accept_content = ['pickle']
 result_serializer = 'pickle'
 timezone = daconfig.get('timezone', 'America/New_York')
 enable_utc = True
+if 'celery processes' in daconfig:
+    worker_concurrency = daconfig['celery processes']